gather all makes into the makes module
